refactor: move leveraged quote function
